Remove redundant condition in hasVotedInFavorOf
As we are already checking that _outcome != OUTCOME_MISSING, if nobody
had voted, _voter would have missing as outcome, so it would return
false without the need of checking the winning outcome.

Besides, it doesn't make sense to ask if some voter voted for missing,
(and if it did it should be equivalent to "has not voted" so it should
return true if voter didn't vote), so moving that part of the
condition to a require.