Remove redundant condition in hasVotedInFavorOf
As we are already checking that _outcome != OUTCOME_MISSING, if nobody
had voted, _voter would have missing as outcome, so it would return
false without the need of checking the winning outcome.

Besides, to ask if some voter voted for missing, should be equivalent
to "has not voted", so it should return true if voter has not voted
yet (no matter what the rest of voters did).