Court: expose external config setter
Fixes #36
Fixes #53
Fixes #90
Fixes #127

It assumes that config has to remain constant throughout all dispute
lifetime. It seems more fair and the logic is simpler.

TODO:

- [ ] Either implement cancel dispute or prevent disputes from being
created in the future (I'd rather do the latter). This is as a
"defensive" mechanism in case the config is changed in between the
dispute is created and it's first draft termId (which is used to get
the config).
- [ ] Fix EIP-170 again.