Court & Voting: compute weight only on commit
It only makes sense to compute juror's weight on commit, not on
reveal. A juror that has not commited won't be able to reveal anyway.

For that reason, order of checks is changed on CRVoting: it makes more
sense and it's more efficient to check conditions locally first and
then if everything is ok go to get weight to Court.